Move terraform/state_{pull,push}.go to statemgmt package
