+++ bundle/statemgmt/state_push_test.go
@@ -1,4 +1,4 @@
-package terraform
+package statemgmt
 
 import (
 	"context"
