+++ cmd/bundle/open.go
@@ -12,6 +12,7 @@ import (
 	"github.com/databricks/cli/bundle/deploy/terraform"
 	"github.com/databricks/cli/bundle/phases"
 	"github.com/databricks/cli/bundle/resources"
+	"github.com/databricks/cli/bundle/statemgmt"
 	"github.com/databricks/cli/cmd/bundle/utils"
 	"github.com/databricks/cli/cmd/root"
 	"github.com/databricks/cli/libs/cmdio"
@@ -87,7 +88,7 @@ func newOpenCommand() *cobra.Command {
 
 		if forcePull || noCache {
 			diags = bundle.ApplySeq(ctx, b,
-				terraform.StatePull(),
+				statemgmt.StatePull(),
 				terraform.Interpolate(),
 				terraform.Write(),
 			)
